zarr storage device created and separated from the upcoming tiff4 storage
