+++ DeviceAdapters/go2scope/go2scope.h
@@ -36,6 +36,10 @@
 
 #define ERR_PARAMETER_ERROR          144001
 #define ERR_INTERNAL						 144002
+#define ERR_ZARR                     140100
+#define ERR_ZARR_SETTINGS            140101
+
+#define ERR_TIFF                     140500
 
 static const char* g_Go2Scope = "Go2Scope";
 static const char* g_MMV1Storage = "MMV1Storage";
@@ -46,44 +50,6 @@ class AcqZarrStorage : public CStorageBase<AcqZarrStorage>
 public:
    AcqZarrStorage();
    virtual ~AcqZarrStorage();
-
-   // Device API
-   // ----------
-   int Initialize();
-   int Shutdown();
-
-   void GetName(char* pszName) const;
-   bool Busy();
-
-   // Storage API
-   // -----------
-   int Create(const char* path, const char* name, int numberOfDimensions, const int shape[], const char* meta, char* handle);
-   int ConfigureDimension(const char* handle, int dimension, const char* name, const char* meaning);
-   int ConfigureCoordinate(const char* handle, int dimension, int coordinate, const char* name);
-   int Close(const char* handle);
-   int Load(const char* path, const char* name, char* handle);
-   int Delete(char* handle);
-   int List(const char* path, char** listOfDatasets, int maxItems, int maxItemLength);
-   int AddImage(const char* handle, unsigned char* pixels, int width, int height, int depth, int coordinates[], int numCoordinates, const char* imageMeta);
-   int GetSummaryMeta(const char* handle, char* meta, int bufSize);
-   int GetImageMeta(const char* handle, int coordinates[], int numCoordinates, char* meta, int bufSize);
-   const unsigned char* GetImage(const char* handle, int coordinates[], int numCoordinates);
-   int GetNumberOfDimensions(const char* handle, int& numDimensions);
-   int GetDimension(const char* handle, int dimension, char* name, int nameLength, char* meaning, int meaningLength);
-   int GetCoordinate(const char* handle, int dimension, int coordinate, char* name, int nameLength);
-
-   // action interface
-   // ----------------
-
-private:
-   bool initialized;
-};
-
-class BigTiffStorage : public CStorageBase<BigTiffStorage>
-{
-public:
-   BigTiffStorage();
-   virtual ~BigTiffStorage();
 
    // Device API
    // ----------
